Ingredients can now be added and removed from the displayed list.
Return button put into Ingredients activity that returns (through a series of steps) an ArrayList back to the MainActivity called ingredientsList. This is the list of ingredients that the user has input. This arrayList can be put into the other activities through a similar method for searching purposes in the yummly api.

Still no way to save/load list. Concentrated on getting the functionality out for the rest of the team to work first.